refactor(b-altoke-local-database): drop unnecessary refs prefix
