+++ bricks/altoke_local_database/reference/_cd___use_drift___drift_local_database/README.md
@@ -1,3 +1,3 @@
-# Altoke Drift Local Database
+# Drift Local Database
 
 A persistent local database built on top of Drift (SQLite).
