+++ bricks/altoke_local_database/reference/_cd___use_drift___drift_local_database/lib/drift_local_database.dart
@@ -1,5 +1,5 @@
 /// A persistent local database built on top of Drift (SQLite).
-library altoke_drift_local_database;
+library drift_local_database;
 
 export 'src/local_database.dart';
 export 'src/local_tasks_dao.dart';
