+++ bricks/altoke_local_database/reference/_cd___use_drift___drift_local_database/lib/src/local_database.dart
@@ -1,7 +1,7 @@
-import 'package:altoke_drift_local_database/src/local_database.drift.dart';
 import 'package:drift/drift.dart';
+import 'package:drift_local_database/src/local_database.drift.dart';
 
-/// {@template altoke_drift_local_database.local_database}
+/// {@template drift_local_database.local_database}
 /// A local database.
 /// {@endtemplate}
 @DriftDatabase(
@@ -10,7 +10,7 @@ import 'package:drift/drift.dart';
   },
 )
 class LocalDatabase extends $LocalDatabase {
-  /// {@macro altoke_drift_local_database.local_database}
+  /// {@macro drift_local_database.local_database}
   LocalDatabase({
     required QueryExecutor queryExecutor,
     required this.schemaVersion,
