+++ bricks/altoke_local_database/reference/_cd___use_drift___drift_local_database/pubspec.yaml
@@ -1,5 +1,5 @@
-name: altoke_isar_local_database
-description: A persistent local database built on top of Isar.
+name: drift_local_database
+description: A persistent local database built on top of Drift (SQLite).
 version: 0.0.1+1
 publish_to: none
 
@@ -16,18 +16,14 @@ dependencies:
   #   path: ../../common/
   #*replace-end*#
   #*w 1v 2> w*#
-  altoke_local_database:
-    path: ../altoke_local_database/
+  local_database:
+    path: ../local_database/
   collection: ^1.19.0
-  isar:
-    hosted: https://pub.isar-community.dev/
-    version: 3.1.8
+  drift: ^2.20.3
   meta: ^1.15.0
 
 dev_dependencies:
   build_runner: ^2.4.12
-  isar_generator:
-    hosted: https://pub.isar-community.dev/
-    version: 3.1.8
+  drift_dev: ^2.20.3
   test: ^1.25.8
   very_good_analysis: ^6.0.0
