+++ bricks/altoke_local_database/reference/_cd___use_isar___isar_local_database/README.md
@@ -1,3 +1,3 @@
-# Altoke Isar Local Database
+# Isar Local Database
 
 A persistent local database built on top of Isar.
