+++ bricks/altoke_local_database/reference/_cd___use_isar___isar_local_database/lib/isar_local_database.dart
@@ -1,5 +1,5 @@
 /// A persistent local database built on top of Isar.
-library altoke_isar_local_database;
+library isar_local_database;
 
 export 'src/local_database_schemas.dart';
 export 'src/local_tasks_dao.dart';
