+++ bricks/altoke_local_database/reference/_cd___use_isar___isar_local_database/lib/src/local_database_schemas.dart
@@ -1,4 +1,4 @@
-import 'package:altoke_isar_local_database/src/task.dart';
+import 'package:isar_local_database/src/task.dart';
 
 /// The local database schemas.
 final localDatabaseSchemas = [
