+++ bricks/altoke_local_database/reference/_cd___use_isar___isar_local_database/pubspec.yaml
@@ -1,5 +1,5 @@
-name: altoke_drift_local_database
-description: A persistent local database built on top of Drift (SQLite).
+name: isar_local_database
+description: A persistent local database built on top of Isar.
 version: 0.0.1+1
 publish_to: none
 
@@ -16,14 +16,18 @@ dependencies:
   #   path: ../../common/
   #*replace-end*#
   #*w 1v 2> w*#
-  altoke_local_database:
-    path: ../altoke_local_database/
+  local_database:
+    path: ../local_database/
   collection: ^1.19.0
-  drift: ^2.20.3
+  isar:
+    hosted: https://pub.isar-community.dev/
+    version: 3.1.8
   meta: ^1.15.0
 
 dev_dependencies:
   build_runner: ^2.4.12
-  drift_dev: ^2.20.3
+  isar_generator:
+    hosted: https://pub.isar-community.dev/
+    version: 3.1.8
   test: ^1.25.8
   very_good_analysis: ^6.0.0
