+++ bricks/altoke_local_database/reference/local_database/lib/src/failures/create_task.dart
@@ -1,19 +1,19 @@
-import 'package:altoke_local_database/altoke_local_database.dart';
+import 'package:local_database/local_database.dart';
 
-/// {@template altoke_local_database.create_task_failure}
+/// {@template local_database.create_task_failure}
 /// An exception that is thrown when a task creation fails.
 /// {@endtemplate}
 sealed class CreateTaskFailure implements Exception {
-  /// {@macro altoke_local_database.create_task_failure}
+  /// {@macro local_database.create_task_failure}
   const CreateTaskFailure();
 }
 
-/// {@template altoke_local_database.create_task_failure_invalid_data}
+/// {@template local_database.create_task_failure_invalid_data}
 /// An exception that is thrown when the data provided for task creation is
 /// invalid.
 /// {@endtemplate}
 class CreateTaskFailureInvalidData extends CreateTaskFailure {
-  /// {@macro altoke_local_database.create_task_failure_invalid_data}
+  /// {@macro local_database.create_task_failure_invalid_data}
   const CreateTaskFailureInvalidData({
     required this.titleValidationErrors,
     required this.complexValidationErrors,
