+++ bricks/altoke_local_database/reference/local_database/lib/src/failures/failures.dart
