+++ bricks/altoke_local_database/reference/local_database/lib/src/models/models.dart
