+++ bricks/altoke_local_database/reference/local_database/pubspec.yaml
@@ -1,4 +1,4 @@
-name: altoke_local_database
+name: local_database
 description: An interface for a persistent local database.
 version: 0.0.1+1
 publish_to: none
